// types/user.ts
export  interface User {
    about?: string;
    active_status_online?: string;
    address?: string;
    alias_mp?: string;
    allow_dm?: string;
    allow_download_files?: string;
    allow_live_streaming_private?: string;
    avatar?: string;
    balance?: string;
    bank?: string;
    birthdate?: string;
    birthdate_changed?: string;
    blocked_countries?: string;
    categories_id?: string;
    city?: string;
    company?: string;
    completed_stripe_onboarding?: string;
    confirmation_code?: string;
    countries_id?: string;
    cover?: string;
    crypto_wallet?: string;
    custom_fee?: string;
    cvu?: string;
    dark_mode?: string;
    date?: string;
    device_token?: string | null;
    discord?: string;
    document_id?: string;
    email?: string;
    email_new_ppv?: string;
    email_new_subscriber?: string;
    email_new_tip?: string;
    facebook?: string;
    featured?: string;
    featured_date?: string | null;
    free_subscription?: string;
    gender?: string;
    github?: string;
    hide_count_subscribers?: string;
    hide_last_seen?: string;
    hide_my_country?: string;
    hide_name?: string;
    hide_profile?: string;
    id: number;
    instagram?: string;
    ip?: string;
    kick?: string;
    language?: string;
    last_login?: string;
    last_seen?: string | null;
    name?: string;
    new_inbox_count?: string;
    new_notifications_count?: string;
    notify_commented_post?: string;
    notify_email_new_post?: string;
    notify_liked_comment?: string;
    notify_liked_post?: string;
    notify_live_streaming?: string;
    notify_mentions?: string;
    notify_new_post?: string;
    notify_new_ppv?: string;
    notify_new_subscriber?: string;
    notify_new_tip?: string;
    oauth_provider?: string;
    oauth_uid?: string;
    payment_gateway?: string;
    payoneer_account?: string;
    paypal_account?: string;
    paystack_authorization_code?: string;
    paystack_card_brand?: string;
    paystack_exp?: string;
    paystack_last4?: string;
    paystack_plan?: string;
    permission?: string;
    permissions?: string;
    pinterest?: string;
    plan?: string;
    pm_last_four?: string | null;
    pm_type?: string | null;
    post_locked?: string;
    posts_privacy?: string;
    price?: string;
    price_live_streaming_private?: string | null;
    price_welcome_message?: string | null;
    profession?: string;
    reddit?: string;
    role?: string;
    send_welcome_message?: string;
    show_my_birthdate?: string;
    snapchat?: string;
    spotify?: string;
    status?: string;
    story?: string;
    stripe_connect_id?: string | null;
    stripe_id?: string | null;
    telegram?: string;
    threads?: string;
    tiktok?: string;
    token?: string;
    trial_ends_at?: string | null;
    twitch?: string;
    twitter?: string;
    two_factor_auth?: string;
    username?: string;
    verified_id?: string;
    vk?: string;
    wallet?: string;
    website?: string;
    welcome_message_new_subs?: string;
    youtube?: string;
    zelle_account?: string;
    zip?: string;
  }
  